Rename it to TicketIgnorer
