+++ src/JiraStatusNotifier/Channel/TicketsByAssigneeClient.php
@@ -4,7 +4,7 @@ declare(strict_types=1);
 
 namespace Chemaclass\JiraStatusNotifier\Channel;
 
-use Chemaclass\JiraStatusNotifier\IgnoreStrategy\IgnoreStrategy;
+use Chemaclass\JiraStatusNotifier\IgnoreStrategy\TicketIgnorer;
 use Chemaclass\JiraStatusNotifier\Jira\Board;
 use Chemaclass\JiraStatusNotifier\Jira\JiraHttpClient;
 use Chemaclass\JiraStatusNotifier\Jira\JqlUrlFactory;
@@ -15,16 +15,16 @@ final class TicketsByAssigneeClient
 
     private JqlUrlFactory $jqlUrlFactory;
 
-    private IgnoreStrategy $ignoreStrategy;
+    private TicketIgnorer $ignoreTickets;
 
     public function __construct(
         JiraHttpClient $jiraClient,
         JqlUrlFactory $jqlUrlFactory,
-        IgnoreStrategy $ignoreUsersStrategy
+        TicketIgnorer $ignoreTickets
     ) {
         $this->jiraClient = $jiraClient;
         $this->jqlUrlFactory = $jqlUrlFactory;
-        $this->ignoreStrategy = $ignoreUsersStrategy;
+        $this->ignoreTickets = $ignoreTickets;
     }
 
     public function fetchFromBoard(Board $board): TicketsByAssignee
@@ -35,7 +35,7 @@ final class TicketsByAssigneeClient
             $tickets = $this->jiraClient->getTickets($this->jqlUrlFactory, $statusName);
 
             foreach ($tickets as $ticket) {
-                if ($this->ignoreStrategy->shouldIgnoreTicket($ticket)) {
+                if ($this->ignoreTickets->shouldIgnore($ticket)) {
                     continue;
                 }
