+++ src/JiraStatusNotifier/IgnoreStrategy/IgnoreStrategy.php
@@ -6,22 +6,7 @@ namespace Chemaclass\JiraStatusNotifier\IgnoreStrategy;
 
 use Chemaclass\JiraStatusNotifier\Jira\ReadModel\JiraTicket;
 
-final class IgnoreStrategy
+interface IgnoreStrategy
 {
-    private IgnoreStrategyInterface $ignoreStrategy;
-
-    public static function byAssigneeKey(string...$assigneeKeys): self
-    {
-        return new self(new IgnoreByAssigneeKey(...$assigneeKeys));
-    }
-
-    private function __construct(IgnoreStrategyInterface $ignoreStrategy)
-    {
-        $this->ignoreStrategy = $ignoreStrategy;
-    }
-
-    public function shouldIgnoreTicket(JiraTicket $ticket): bool
-    {
-        return $this->ignoreStrategy->shouldIgnore($ticket);
-    }
+    public function shouldIgnore(JiraTicket $ticket): bool;
 }
