+++ src/JiraStatusNotifier/JiraConnector.php
@@ -7,7 +7,7 @@ namespace Chemaclass\JiraStatusNotifier;
 use Chemaclass\JiraStatusNotifier\Channel\ChannelInterface;
 use Chemaclass\JiraStatusNotifier\Channel\ChannelResult;
 use Chemaclass\JiraStatusNotifier\Channel\TicketsByAssigneeClient;
-use Chemaclass\JiraStatusNotifier\IgnoreStrategy\IgnoreStrategy;
+use Chemaclass\JiraStatusNotifier\IgnoreStrategy\TicketIgnorer;
 use Chemaclass\JiraStatusNotifier\IO\JiraConnectorInput;
 use Chemaclass\JiraStatusNotifier\Jira\Board;
 use Chemaclass\JiraStatusNotifier\Jira\JiraHttpClient;
@@ -42,7 +42,7 @@ final class JiraConnector
         $ticketsByAssignee = (new TicketsByAssigneeClient(
             $this->jiraHttpClient,
             new JqlUrlFactory($jiraBoard, JqlUrlBuilder::inOpenSprints($company)),
-            IgnoreStrategy::byAssigneeKey(...$input->jiraUsersToIgnore())
+            TicketIgnorer::byAssigneeKey(...$input->jiraUsersToIgnore())
         ))->fetchFromBoard($jiraBoard);
 
         foreach ($this->channels as $channel) {
