+++ tests/JiraStatusNotifierTest/Unit/IgnoreStrategy/IgnoreStrategyTest.php
@@ -4,7 +4,7 @@ declare(strict_types=1);
 
 namespace Chemaclass\JiraStatusNotifierTests\Unit\IgnoreStrategy;
 
-use Chemaclass\JiraStatusNotifier\IgnoreStrategy\IgnoreStrategy;
+use Chemaclass\JiraStatusNotifier\IgnoreStrategy\TicketIgnorer;
 use Chemaclass\JiraStatusNotifier\Jira\ReadModel\Assignee;
 use Chemaclass\JiraStatusNotifier\Jira\ReadModel\JiraTicket;
 use Chemaclass\JiraStatusNotifier\Jira\ReadModel\TicketStatus;
@@ -16,15 +16,15 @@ final class IgnoreStrategyTest extends TestCase
     /** @test */
     public function shouldIgnore(): void
     {
-        $ignoreStrategy = IgnoreStrategy::byAssigneeKey('key1', 'key2');
-        self::assertTrue($ignoreStrategy->shouldIgnoreTicket($this->newTicket('key1')));
+        $ignoreStrategy = TicketIgnorer::byAssigneeKey('key1', 'key2');
+        self::assertTrue($ignoreStrategy->shouldIgnore($this->newTicket('key1')));
     }
 
     /** @test */
     public function shouldNotIgnore(): void
     {
-        $policy = IgnoreStrategy::byAssigneeKey('key1', 'key2');
-        self::assertFalse($policy->shouldIgnoreTicket($this->newTicket('key3')));
+        $policy = TicketIgnorer::byAssigneeKey('key1', 'key2');
+        self::assertFalse($policy->shouldIgnore($this->newTicket('key3')));
     }
 
     private function newTicket(string $key): JiraTicket
